cpp: Add .clang-format file
It's convenient to have a code formatter while developing, so that it
can fix up the syntax on save.

Many IDEs / editor tools rely on the presence of a `.clang-format` file
to enable autoformating files on save. Adding this file makes it easier
for candidates to clang-format, while not hurting candidates who haven't
set up their development environment to use this.

I've added the `.clang-format` file that we use in
<https://github.com/sorbet/sorbet> and reformatted the starter code
according to it.