java: Upgrade okhttp to latest 4.x (4.9.0)
OkHttp's site now has a 4.x version, and the code snippet on their main
page uses this version. The version we had before this is significantly
different in that MediaType.get is not defined at all, even though both
3.x and 4.x documentation say it exists.

I upgraded to 4.x rather than the latest 3.x because I don't see why we
should be on an older version. I confirmed that we can still run the code
we need for the exercise just fine.